SignUp fixes + Clean code

+++ scripts/login-signup.js
@@ -4,7 +4,6 @@ if(uniqueUserId==null) uniqueUserId = 1010;
 
 // Function invoked on User Sign Up
 function signUp(){
-    console.log("Testing scripting working or not")
     let allUsers = JSON.parse(localStorage.getItem("allUsers"));
     if(allUsers==null) allUsers=[];
     var newUsertoAdd = {
@@ -14,6 +13,13 @@ function signUp(){
         userEmail: document.querySelector("#email").value,
         userPass: document.querySelector("#password").value
     };
+    if(allUsers.length!=0 && userExists(newUsertoAdd.userEmail)){
+        document.querySelector("#sign-up-id").textContent = "User already exists...";
+        setTimeout(()=>{
+            document.querySelector("#sign-up-id").textContent = "Sign Up";
+        },1200)
+        return;
+    }
     allUsers.push(newUsertoAdd);
     localStorage.setItem("loggedDetails",JSON.stringify([{isLogged:true},newUsertoAdd]))
     localStorage.setItem("uniqueUserCounter",JSON.stringify(++uniqueUserId));
@@ -33,17 +39,24 @@ function signUp(){
     }
     localStorage.removeItem("userComingfrom");
 }
-//document.querySelector("#sign-up-id").addEventListener('click',signUp);
+
+function userExists(newUserEmail){
+    let allUsers = JSON.parse(localStorage.getItem("allUsers"));
+    for(var i=0; i<allUsers.length; i++){
+        if(allUsers[i].userEmail==newUserEmail) return true;
+    }
+    return false;
+}
 
 let userIdforLogin = null;
-
+let loginButton = document.querySelector("#login-btn-id");
 // Function Invoked on User Login
 function logIn(){
     let mailEntered = document.querySelector("#email").value;
     let passEntered = document.querySelector("#password").value;
     if(checkUserDetails(mailEntered,passEntered)){
         localStorage.setItem("loggedDetails",JSON.stringify([{isLogged:true},userIdforLogin]));
-        document.querySelector("#login-btn-id").textContent = "Login Success...";
+        loginButton.textContent = "Login Success...";
         let pageFrom = JSON.parse(localStorage.getItem("userComingfrom"));
         if(pageFrom=="projectDetailsPage"){
             setTimeout(()=>{
@@ -56,8 +69,10 @@ function logIn(){
         }
         localStorage.removeItem("userComingfrom");
     }else{
-        console.log("User Name not correct");
-        alert("User Id Password does not match !")
+        loginButton.textContent = "Invalid Credentials...";
+        setTimeout(()=>{
+            loginButton.textContent = "Login";
+        },1000)
     }
 }
 
@@ -74,4 +89,3 @@ function checkUserDetails(mailEntered,passEntered){
     }
     return false;
 }
-//document.querySelector("#login-btn-id").addEventListener("click",logIn);
